refactor/#104: envConfig, endpoints로 파일 분리
